requirements.txt
basemap==1.3.6
basemap-data==1.3.2
certifi==2022.9.24
contourpy==1.0.6
cycler==0.11.0
fonttools==4.38.0
kiwisolver==1.4.4
matplotlib==3.6.2
numpy==1.23.5
packaging==21.3
Pillow==9.3.0
pyparsing==3.0.9
pyproj==3.4.0
PyQt5==5.15.7
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
pyshp==2.3.1
python-dateutil==2.8.2
scipy==1.9.3
six==1.16.0
